// src/types/interfaces.tsx
// export interface Coordinates {
//   lat: number;
//   lon: number;
// }

// export interface GeoBoundingBox {
//   top?: number;
//   left?: number;
//   bottom?: number;
//   right?: number;
//   bottom_left?: Coordinates;
//   top_right?: Coordinates;
//   bottom_right?: Coordinates;
//   top_left?: Coordinates;
// }
// export interface Dog {
//   id: string;
//   img: string;
//   name: string;
//   age: number;
//   zip_code: string;
//   breed: string;
// }

// export interface SearchResponse {
//   resultIds: string[];
//   total: number;
//   next: string | null;
//   prev: string | null;
// }

// export interface MatchResponse {
//   match: string;
// }


// export interface Location {
//   zip_code: string;
//   latitude: number;
//   longitude: number;
//   city: string;
//   state: string;
//   county: string;
// }

// export interface LocationSearchParams {
//   city?: string;
//   states?: string[];
//   geoBoundingBox?: GeoBoundingBox;
//   size?: number;
//   from?: number;
// }

// export interface LocationSearchResponse {
//   results: Location[];
//   total: number;
// }

export interface Coordinates {
  lat: number;
  lon: number;
}

export interface GeoBoundingBox {
  top?: number;
  left?: number;
  bottom?: number;
  right?: number;
  bottom_left?: Coordinates;
  top_right?: Coordinates;
  bottom_right?: Coordinates;
  top_left?: Coordinates;
}

export interface Location {
  zip_code: string;
  latitude: number;
  longitude: number;
  city: string;
  state: string;
  county: string;
}

export interface LocationSearchParams {
  city?: string;
  states?: string[];
  geoBoundingBox?: GeoBoundingBox;
  size?: number;
  from?: number;
}

export interface LocationSearchResponse {
  results: Location[];
  total: number;
}

export interface Dog {
  id: string;
  img: string;
  name: string;
  age: number;
  zip_code: string;
  breed: string;
}

export interface SearchResponse {
  resultIds: string[];
  total: number;
  next: string | null;
  prev: string | null;
}

export interface MatchResponse {
  match: string;
}